refactor(http): move auth mod to handlers::common
The `auth` mod does not contains a handler.
It only contains auth error and funtion to ma the error into responses.
