+++ src/http/axum_implementation/handlers/announce.rs
@@ -11,7 +11,7 @@ use super::common::peer_ip;
 use crate::http::axum_implementation::extractors::announce_request::ExtractRequest;
 use crate::http::axum_implementation::extractors::key::Extract;
 use crate::http::axum_implementation::extractors::remote_client_ip::RemoteClientIp;
-use crate::http::axum_implementation::handlers::auth;
+use crate::http::axum_implementation::handlers::common::auth;
 use crate::http::axum_implementation::requests::announce::{Announce, Compact, Event};
 use crate::http::axum_implementation::responses::{self, announce};
 use crate::http::axum_implementation::services;
