+++ src/http/axum_implementation/handlers/common/mod.rs
@@ -1,1 +1,2 @@
 pub mod peer_ip;
+pub mod auth;
